+++ packages/react-theme-provider/etc/react-theme-provider.api.md
@@ -8,15 +8,10 @@ import { ColorTokenSet } from '@fluentui/theme';
 import { ICustomizerContext } from '@fluentui/utilities';
 import { IFontFace } from '@fluentui/merge-styles';
 import { IKeyframes } from '@fluentui/merge-styles';
-import { IRawFontStyle } from '@fluentui/merge-styles';
-import { IRawStyle } from '@fluentui/merge-styles';
-import { IStyle } from '@fluentui/merge-styles';
-import { IStyleFunctionOrObject } from '@fluentui/merge-styles';
 import { PartialTheme } from '@fluentui/theme';
 import * as React from 'react';
 import { Theme } from '@fluentui/theme';
 import { TokenSetType } from '@fluentui/theme';
-import { Variants } from '@fluentui/theme';
 
 // @public
 export const applyClasses: <TState extends {}>(state: TState, classMap: Record<string, string>) => void;
@@ -27,37 +22,8 @@ export type FontFace = IFontFace;
 // @public (undocumented)
 export const getStyleFromPropsAndOptions: <TProps extends StyleProps<import("@fluentui/theme").ColorTokenSet>, TOptions extends StyleOptions<TProps>>(props: TProps, options: TOptions, prefix?: string | undefined) => React.CSSProperties;
 
-export { IRawFontStyle }
-
-export { IRawStyle }
-
-export { IStyle }
-
-export { IStyleFunctionOrObject }
-
 // @public (undocumented)
 export type KeyFrames = IKeyframes;
-
-// @public
-export const makeClasses: <TState extends {}>(styleOrFunction: Record<string, IStyle> | ((theme: Theme) => Record<string, IStyle>)) => (state: TState, options?: UseStylesOptions | undefined) => void;
-
-// @public
-export function makeStyles<TStyleSet extends {
-    [key: string]: IStyle;
-}>(styleOrFunction: TStyleSet | ((theme: Theme) => TStyleSet)): (options?: UseStylesOptions) => {
-    [key in keyof TStyleSet]: string;
-};
-
-// @public
-export const makeVariantClasses: <TState = {}, TVariants = Record<string, any>>(options: MakeVariantClassesOptions<TVariants>) => (state: TState, options?: import("./makeStyles").UseStylesOptions | undefined) => void;
-
-// @public
-export type MakeVariantClassesOptions<TVariants = Variants> = {
-    name?: string;
-    prefix?: string;
-    styles?: Record<string, IStyle> | ((theme: Theme) => Record<string, IStyle>);
-    variants?: TVariants | ((theme: Theme) => TVariants);
-};
 
 // @public (undocumented)
 export const MergeStylesProvider: ({ children }: {
@@ -138,12 +104,6 @@ export const useInlineTokens: (draftState: {
 
 // @public (undocumented)
 export const useStyleRenderer: () => StyleRenderer;
-
-// @public
-export type UseStylesOptions = {
-    theme?: Theme;
-    renderer?: StyleRenderer;
-};
 
 // @public
 export const useTheme: () => Theme;
