+++ packages/react-theme-provider/src/makeStyles.compat.test.tsx
@@ -6,7 +6,7 @@ import { Stylesheet, InjectionMode } from '@fluentui/merge-styles';
 import { safeMount } from '@fluentui/test-utilities';
 import { mount, ReactWrapper } from 'enzyme';
 import { act } from 'react-dom/test-utils';
-import { makeStyles } from './makeStyles';
+import { makeStyles } from './makeStyles.compat';
 import { MergeStylesProvider } from './styleRenderers/mergeStylesRenderer';
 import { ThemeProvider } from './ThemeProvider';
