+++ packages/react-theme-provider/src/makeStyles.compat.ts
